backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langgraph==0.0.65
langchain==0.0.335
httpx==0.25.0
pydantic==2.4.2
python-dotenv==1.0.0
aiofiles==23.2.0
google-generativeai==0.3.2
